Refactor and add ControlBarViewModel
